:bug: fix rename package name in Student-API
