--- API/Student-API/src/main/java/tw/waterball/judgegirl/problemapi/clients/StudentServiceDriver.java
@@ -1,12 +1,0 @@
-package tw.waterball.judgegirl.problemapi.clients;
-
-import tw.waterball.judgegirl.entities.Student;
-
-import java.util.List;
-
-public interface StudentServiceDriver {
-
-    List<Student> getStudentsByIds(List<Integer> ids);
-
-    List<Student> getStudentsByEmails(List<String> emails);
-}
