+++ API/Student-API/src/main/java/tw/waterball/judgegirl/studentapi/clients/StudentServiceDriver.java
@@ -1,0 +1,12 @@
+package tw.waterball.judgegirl.studentapi.clients;
+
+import tw.waterball.judgegirl.entities.Student;
+
+import java.util.List;
+
+public interface StudentServiceDriver {
+
+    List<Student> getStudentsByIds(List<Integer> ids);
+
+    List<Student> getStudentsByEmails(List<String> emails);
+}
