+++ API/Student-API/src/main/java/tw/waterball/judgegirl/studentapi/clients/view/GroupView.java
@@ -1,4 +1,4 @@
-package tw.waterball.judgegirl.problemapi.clients.view;
+package tw.waterball.judgegirl.studentapi.clients.view;
 
 import lombok.*;
 import tw.waterball.judgegirl.entities.Group;
