+++ API/Student-API/src/main/java/tw/waterball/judgegirl/studentapi/clients/view/StudentView.java
@@ -11,7 +11,7 @@
  *   limitations under the License.
  */
 
-package tw.waterball.judgegirl.problemapi.clients.view;
+package tw.waterball.judgegirl.studentapi.clients.view;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
