+++ Exam-Service/src/main/java/tw/waterball/judgegirl/examservice/domain/usecases/exam/GetExamineesUseCase.java
@@ -5,7 +5,7 @@ import tw.waterball.judgegirl.entities.Student;
 import tw.waterball.judgegirl.entities.exam.Exam;
 import tw.waterball.judgegirl.entities.exam.Examinee;
 import tw.waterball.judgegirl.examservice.domain.repositories.ExamRepository;
-import tw.waterball.judgegirl.problemapi.clients.StudentServiceDriver;
+import tw.waterball.judgegirl.studentapi.clients.StudentServiceDriver;
 
 import javax.inject.Named;
 import java.util.List;
