+++ Judger/src/main/java/tw/waterball/judgegirl/judger/CCJudger.java
@@ -35,8 +35,8 @@ import tw.waterball.judgegirl.judger.infra.testexecutor.TestcaseExecutor;
 import tw.waterball.judgegirl.judger.infra.testexecutor.TestcaseExecutorFactory;
 import tw.waterball.judgegirl.judger.layout.*;
 import tw.waterball.judgegirl.plugins.api.JudgeGirlPluginLocator;
-import tw.waterball.judgegirl.problemapi.clients.ProblemServiceDriver;
 import tw.waterball.judgegirl.problemapi.views.ProblemView;
+import tw.waterball.judgegirl.studentapi.clients.ProblemServiceDriver;
 import tw.waterball.judgegirl.submissionapi.clients.SubmissionServiceDriver;
 import tw.waterball.judgegirl.submissionapi.clients.VerdictPublisher;
 import tw.waterball.judgegirl.submissionapi.views.SubmissionView;
