+++ Judger/src/main/java/tw/waterball/judgegirl/judger/DefaultCCJudgerFactory.java
@@ -31,9 +31,9 @@ import tw.waterball.judgegirl.judgerapi.env.JudgerEnvVariables;
 import tw.waterball.judgegirl.plugins.Plugins;
 import tw.waterball.judgegirl.plugins.api.JudgeGirlPlugin;
 import tw.waterball.judgegirl.plugins.api.PresetJudgeGirlPluginLocator;
-import tw.waterball.judgegirl.problemapi.clients.ProblemApiClient;
-import tw.waterball.judgegirl.problemapi.clients.ProblemServiceDriver;
 import tw.waterball.judgegirl.springboot.amqp.AmqpVerdictPublisher;
+import tw.waterball.judgegirl.studentapi.clients.ProblemApiClient;
+import tw.waterball.judgegirl.studentapi.clients.ProblemServiceDriver;
 import tw.waterball.judgegirl.submissionapi.clients.SubmissionApiClient;
 import tw.waterball.judgegirl.submissionapi.clients.SubmissionServiceDriver;
 import tw.waterball.judgegirl.submissionapi.clients.VerdictPublisher;
