+++ Spring-Boot/Spring-Boot-Exam/src/test/java/tw/waterball/judgegirl/springboot/exam/controllers/ExamControllerTest.java
@@ -33,8 +33,6 @@ import tw.waterball.judgegirl.examservice.domain.repositories.ExamRepository;
 import tw.waterball.judgegirl.examservice.domain.usecases.exam.CreateExamUseCase;
 import tw.waterball.judgegirl.examservice.domain.usecases.exam.CreateQuestionUseCase;
 import tw.waterball.judgegirl.examservice.domain.usecases.exam.UpdateExamUseCase;
-import tw.waterball.judgegirl.problemapi.clients.FakeProblemServiceDriver;
-import tw.waterball.judgegirl.problemapi.clients.FakeStudentServiceDriver;
 import tw.waterball.judgegirl.problemapi.views.ProblemView;
 import tw.waterball.judgegirl.springboot.exam.SpringBootExamApplication;
 import tw.waterball.judgegirl.springboot.exam.handler.VerdictIssuedEventHandler;
@@ -43,6 +41,8 @@ import tw.waterball.judgegirl.springboot.exam.view.ExamHome;
 import tw.waterball.judgegirl.springboot.exam.view.ExamView;
 import tw.waterball.judgegirl.springboot.exam.view.QuestionView;
 import tw.waterball.judgegirl.springboot.profiles.Profiles;
+import tw.waterball.judgegirl.studentapi.clients.FakeProblemServiceDriver;
+import tw.waterball.judgegirl.studentapi.clients.FakeStudentServiceDriver;
 import tw.waterball.judgegirl.submissionapi.clients.SubmissionServiceDriver;
 import tw.waterball.judgegirl.submissionapi.clients.VerdictPublisher;
 import tw.waterball.judgegirl.submissionapi.views.SubmissionView;
