+++ Spring-Boot/Spring-Boot-Student/src/main/java/tw/waterball/judgegirl/springboot/student/presenters/SignUpPresenter.java
@@ -15,10 +15,10 @@ package tw.waterball.judgegirl.springboot.student.presenters;
 
 import org.springframework.stereotype.Component;
 import tw.waterball.judgegirl.entities.Student;
-import tw.waterball.judgegirl.problemapi.clients.view.StudentView;
+import tw.waterball.judgegirl.studentapi.clients.view.StudentView;
 import tw.waterball.judgegirl.studentservice.domain.usecases.student.SignUpUseCase;
 
-import static tw.waterball.judgegirl.problemapi.clients.view.StudentView.toViewModel;
+import static tw.waterball.judgegirl.studentapi.clients.view.StudentView.toViewModel;
 
 /**
  * @author chaoyulee chaoyu2330@gmail.com
